pyproject.toml
[tool.poetry]
name = "course-paper-with-sql"
version = "0.1.0"
description = ""
authors = ["denishudyakov <denis.hudyakov27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
sqlalchemy = "^2.0.27"


[tool.poetry.group.lint.dependencies]
black = "^24.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
